src/app/components/welcome/welcome.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';

import { WelcomeComponent } from './welcome.component';

describe('WelcomeComponent', () => {
  let component: WelcomeComponent;
  let fixture: ComponentFixture<WelcomeComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [ WelcomeComponent ]
    })
      .compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(WelcomeComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should render welcome message', () => {
    const fixture = TestBed.createComponent(WelcomeComponent);
    fixture.detectChanges();
    const compiled = fixture.nativeElement as HTMLElement;
    expect(compiled.querySelector('#title')?.textContent).toContain('Bienvenidos a mi portfolio');
    expect(compiled.querySelector('#description')?.textContent).toContain('Algunos de mis recorridos de vida');
    expect(compiled.querySelector('#book')?.textContent).toContain('Versión alternativa: diario de viaje');
    expect(compiled.querySelector('#map')?.textContent).toContain('Navegar los itinerarios en el mapa');

  });
});
